Parse envs
