+++ packages/api/.env.sample
@@ -1,11 +1,5 @@
 NODE_ENV=development
 PORT=5000
 
-NODE_ENV=development
-PORT=5000
-
-
 POSTGRES_PRISMA_URL=
-# Direct connection to the database. Used for migrations.
 POSTGRES_URL_NON_POOLING=
-POSTGRES_PASSWORD=
